start implementing game state serialization
