+++ BattleFriends/BattleFriends.vcxproj.filters
@@ -36,6 +36,9 @@
     <ClCompile Include="src\minimap.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="src\save_game_state.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\include\BF.h">
@@ -66,6 +69,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="src\include\loadResource.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\include\save_game_state.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
