+++ BattleFriends/src/include/save_game_state.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace BF
+{
+	bool save_game_state(unsigned char** buffer, int* len, int* checksum, int frame);
+}
